refactor: split report/nodes.py into nodes/ subpackage (BTND-07-001)
Converts vultron/core/behaviors/report/nodes.py (1063 lines, 15 classes)
into a nodes/ subpackage grouped by semantic concern per BTND-07-001/002:

- conditions.py: 7 Check*/Evaluate*/Ensure* condition nodes
- rm_transitions.py: TransitionRMtoValid, TransitionRMtoInvalid
- case_creation.py: CreateCaseNode, CreateCaseActivity + private helpers
- participant.py: TransitionParticipantRMtoAccepted/Deferred
- emit.py: EmitEngageCaseActivity, EmitDeferCaseActivity
- __init__.py: re-exports all public names for backward compatibility

Mirror the split in test/core/behaviors/report/nodes/ with one
test_<submodule>.py per submodule. All existing import paths unchanged.

Closes #738
